requirements.txt
# requirements.txt
Flask
gunicorn
Pillow
python-dotenv
requests
beautifulsoup4
feedparser
google-api-python-client
google-auth-oauthlib